@@ -18,12 +18,14 @@
     "firebase-functions": "^3.11.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^3.9.1",
-    "@typescript-eslint/parser": "^3.8.0",
+    "@types/express": "^4.17.9",
+    "@types/express-serve-static-core": "^4.17.14",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
     "eslint": "^7.6.0",
     "eslint-plugin-import": "^2.22.0",
-    "typescript": "^3.8.0",
-    "firebase-functions-test": "^0.2.0"
+    "firebase-functions-test": "^0.2.0",
+    "typescript": "^4.1.2"
   },
   "private": true
 }
